Chapter 1 rewritten
"Jam Mode" has also been implmented, so to represent the original Hack Jam
entry this was originally intended to be. That ideally will actually end
after killing Alear in Ch5 as opposed to an End of Demo screen at Ch6. If
I can even make it to Ch6: that might end at Ch5 too!