Cleaned up Chapter 1 a bit
Fixed the palette meltdowns in the Jam Mode opening, just gave Marth money
at the start to avoid having to burn a turn selling the blue gem, moved the
save point so units can't spawn on top of it, and gave Marth a member card to
get rid of one of the many "has item in story but not gameplay" moments.
Now I just need to scatter some secret shops around!