Fixed Power Ring, properly implemented Spirit Dust
By default Spirit Dust replaces the GBAFE Energy Ring. So it was giving magic
all along. So I had to go and fix that. I wasted a lot of time looking in the
IER folder looking for the table, but turns out it was a CSV.

While I had the item icon table open, I made some more placeholder icons too.