Added Ice Breath animation, changed Mage Female animation
Lilina gets main character priveliges so all the mage girls look like her
instead of Nino now. Sure there's fancier Lilinas in the repo, but the vanilla
sprite is mage female, so others still work with it.
Also, there's a whole lot of transformation files I don't need: Nils is
mentioned but not seen, and Ninian is dead!