Started extending item names
Because I'm not a wizard, I spent a few hours trying to trace a magic constant
in the decomp to the byte it came from in the rom. Not very smart - the real
solution would be to run the decomp, change what needs changing, and then see
what you need to change. But I don't feel like installing a million things.
Lousy excuse, I know...