Extended several item names
Eliminated the abbreviations on "Shield of Seals" and "Thunder Sword" (the
latter just barely fits in eight tiles!) and added "Staff" to the end of the
staves in line with the Japanese item names in the SNES era.
(I won't change the magic though - nobody says Book of X.)
Fixed (by my definition) the spelling of "Armourslayer" too.