Fixed the bug with Page 3's item icons
Well, actually Sme fixed the bug because I can't compile C code, apparently.
That was a bit of a saga involving trying to edit the .lyn.event file (don't)
and posting in the wrong channel a few times on the FEU discord server.
Now I can clear out the redundant item icon slots and put in something useful.