Polishing after testing
The weird glitched coastlines are fixed, the bug stopping dogs running indoors
(that's a rule break, as far as I know), and the bad layout stopping you from
actually finishing the map in reasonable time by the armoury has been fixed.
A few numbers have been tweaked to make the chapter a bit easier on normal.
Hard is untested and not recommended: stats go up just enough that keeping
Enteh alive is as good as impossible.